[CORE-1395] Move PostgreSQL collection v2.5.0 migration tests (#8501)
* Move PostgreSQL collection v2.5.0 migration tests.
